Rawr.DPSDK: Completely redesigned how abilities are handled; completely replaced rotation system with non-user-definable blood/frost/unholy priority systems; fixed numerous calculation errors; improved KM and Rime handling; death rune talents now weighted; lots and lots of changes, hopefully nothing is (too) broken. Performance improvements will be coming in the future, as these changes have had a relatively large reduction in calculation speed. Rawr2 DPSDK options pane is now largely obsolete, but it's being ignored for now. Rawr3 options pane is also obsolete, and an update for it will be coming soon.
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS03]$/Rawr;C45485

+++ Rawr.DPSDK/DRW.cs
@@ -83,7 +83,7 @@ namespace Rawr.DPSDK
                 float dpsMHBCB = 0f;
                 if (combatTable.MH.damage != 0)
                 {
-                    float MHBCBDmg = combatTable.MH.damage * (.25f + .125f * calcOpts.rotation.AvgDiseaseMult);
+                    float MHBCBDmg = (float)(combatTable.MH.damage * (.25f + .125f * calcOpts.rotation.AvgDiseaseMult));
                     dpsMHBCB = MHBCBDmg / combatTable.MH.hastedSpeed;
                 }
                 dpsBCB = dpsMHBCB;
